refactor: Rm a lot of genetic code
