+++ AIAgent/ml/training/wrapper.py
@@ -1,10 +1,10 @@
 import copy
+
 import torch
 from common.game import GameState
-from ml.training.dataset import convert_input_to_tensor
-from ml.model_wrappers.protocols import Predictor
-
 from ml.predict import predict_state_with_dict
+from ml.protocols import Predictor
+from ml.training.dataset import convert_input_to_tensor
 
 
 class TrainingModelWrapper(Predictor):
